// web/src/api/generated/models/friend-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 好友信息响应
 */
export interface FriendResponse {
    /**
     * 用户ID
     */
    'userId'?: number;
    /**
     * 用户账号
     */
    'userAccount'?: string;
    /**
     * 用户名
     */
    'userName'?: string;
    /**
     * 用户头像
     */
    'userAvatar'?: string;
    /**
     * 个人简介
     */
    'userProfile'?: string;
    /**
     * 成为好友时间
     */
    'friendSince'?: string;
}