// web/src/api/generated/models/friend-request-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 好友申请响应
 */
export interface FriendRequestResponse {
    /**
     * 申请ID
     */
    'id'?: number;
    /**
     * 发送者ID
     */
    'senderId'?: number;
    /**
     * 发送者用户名
     */
    'senderName'?: string;
    /**
     * 发送者头像
     */
    'senderAvatar'?: string;
    /**
     * 接收者ID
     */
    'receiverId'?: number;
    /**
     * 接收者用户名
     */
    'receiverName'?: string;
    /**
     * 接收者头像
     */
    'receiverAvatar'?: string;
    /**
     * 申请状态：pending/accepted/rejected
     */
    'status'?: string;
    /**
     * 申请消息
     */
    'message'?: string;
    /**
     * 创建时间
     */
    'createTime'?: string;
}