// web/src/api/generated/models/search-user-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 搜索用户响应
 */
export interface SearchUserResponse {
    /**
     * 用户ID
     */
    'userId'?: number;
    /**
     * 用户账号
     */
    'userAccount'?: string;
    /**
     * 用户名
     */
    'userName'?: string;
    /**
     * 用户头像
     */
    'userAvatar'?: string;
    /**
     * 个人简介
     */
    'userProfile'?: string;
    /**
     * 是否已是好友
     */
    'isFriend'?: boolean;
    /**
     * 是否有待处理的申请
     */
    'hasPendingRequest'?: boolean;
}